+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ## [unreleased]
 
- * External Redis Messaging Provider: Fixed a bug in the health probe which led to the situation that the server never reach ready state
+ * MongoDB Storage Provider: Fixed a potential problem in the health probe which could lead to the situation that the server never reaches ready state
+ * External Redis Messaging Provider: Fixed a bug in the health probe which could lead to the situation that the server never reaches ready state
  * Rest Proxy Demo App: Replaced legacy *spacexdata.com* API
 
 ## 2.4.0 (June 9, 2023)